requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-environ==0.9.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.0
phonenumbers==8.13.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
twilio==7.15.2
uritemplate==4.1.1
urllib3==1.26.12